# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pharma/ask:
    post:
      operationId: askPharmaAssistant
      tags: [pharma]
      summary: Answer a pharmacy or drug-related question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmaQuestionInput"
      responses:
        "200":
          description: Educational assistant answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmaAnswer"
        "400":
          description: Invalid question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmaError"
        "502":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmaError"
        "503":
          description: AI provider is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmaError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PharmaQuestionInput:
      type: object
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 1000
        mode:
          type: string
          enum:
            - ask
            - drug-profile
            - study-session
            - quiz-generation
        context:
          type: string
          maxLength: 12000
      required:
        - question
    PharmaAnswer:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer
    PharmaError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...